(PC-14474) refactor(Booking): extract TicketBody component from BookingDetailsTicketContent
